Added screenshots and updated README
