+++ .env
@@ -1,0 +1,2 @@
+GOOGLE_API_KEY= 'AIzaSyCzN7cYMZP0IncvlDmw_GbetdgkqT5krMU'
+MYSQL_PASSWORD= 'root'
