+++ pyproject.toml
@@ -16,11 +16,11 @@ dependencies = [
     "tiktoken (>=0.9.0,<0.10.0)",
     "faiss-cpu (>=1.10.0,<2.0.0)",
     "protobuf (>=6.30.2,<7.0.0)",
-    "langchain-experimental (>=0.3.4,<0.4.0)",
     "mysql-connector-python (>=9.2.0,<10.0.0)",
     "pymysql (>=1.1.1,<2.0.0)",
     "sentence-transformers (>=4.0.1,<5.0.0)",
     "cryptography (>=44.0.2,<45.0.0)",
+    "langchain-experimental (>=0.3.4,<0.4.0)",
 ]
